tyck: cleanup when exiting from a scope
